[ADD] Add automation script for db report sync.

+++ gbs_ogl_cbs_interface/models/glif_process.py
@@ -11,6 +11,7 @@ from odoo import exceptions, models, fields, api, _, tools
 from odoo.exceptions import ValidationError
 
 TIME_FORMAT = "%Y-%m-%d %H:%M:%S"
+DATE_FORMAT = "%Y-%m-%d"
 
 _logger = logging.getLogger(__name__)
 try:
@@ -981,16 +982,18 @@ class ServerFileSuccess(models.Model):
     _name = 'server.file.success'
     _description = "GLIF File Success"
     _inherit = ["mail.thread", "ir.needaction_mixin"]
-    _order = 'name desc'
+    _order = 'glif_date desc'
 
     name = fields.Char(string='Name', compute='_compute_name', store=True)
-    start_date = fields.Datetime(string='Start Datetime', required=True)
-    stop_date = fields.Datetime(string='Stop Datetime', required=True)
+    glif_date = fields.Date(string='GLIF Date', compute="_compute_glif_date", store=True)
+    start_date = fields.Datetime(string='Processing Start', required=True)
+    stop_date = fields.Datetime(string='Processing End', required=True)
     file_name = fields.Char(string='File Path', required=True)
     upload_file = fields.Binary(string="Upload File", attachment=True)
     time = fields.Text(string='Time', compute="_compute_time")
     move_id = fields.Many2one('account.move', string="Journal")
     status = fields.Boolean(default=False, string='Status')
+    state = fields.Selection([('pending', 'Pending'), ('success', 'Success')], default='pending', string='Report Sync')
 
     @api.depends('start_date', 'stop_date')
     def _compute_time(self):
@@ -998,15 +1001,23 @@ class ServerFileSuccess(models.Model):
             diff = dt.strptime(rec.stop_date, TIME_FORMAT) - dt.strptime(rec.start_date, TIME_FORMAT)
             rec.time = str(diff)
 
+    @api.depends('name')
+    def _compute_glif_date(self):
+        for rec in self:
+            if rec.name:
+                date = dt.strptime(str(rec.name[4:8] + "-" + rec.name[0:2] + "-" + rec.name[2:4]), DATE_FORMAT)
+                rec.glif_date = date
+
     @api.multi
     def unlink_glif(self):
-        unlink = "UPDATE account_move SET state='draft' WHERE id=%s" % self.move_id.id
-        self.env.cr.execute(unlink)
-        self._cr.commit()
-        if not self.env.cr.execute(unlink):
-            self.move_id.unlink()
-            self.unlink()
+        if self.move_id:
+            unlink = "UPDATE account_move SET state='draft' WHERE id=%s" % self.move_id.id
+            self.env.cr.execute(unlink)
             self._cr.commit()
+            if not self.env.cr.execute(unlink):
+                self.move_id.unlink()
+        self.unlink()
+        self._cr.commit()
 
 
 class GenerateCBSJournalSuccess(models.Model):
